🔨 Moved tauri commands into its own file, moved existing struct into types module
