+++ src-tauri/src/types/mod.rs
@@ -1,0 +1,2 @@
+pub mod types;
+pub use types::*;
