+++ src-tauri/src/types/types.rs
@@ -1,0 +1,7 @@
+use std::sync::Arc;
+use discordipc::InnerClient;
+
+pub struct AppData {
+    pub drpc_client: Arc<discordipc::Client<Arc<InnerClient>>>,
+    pub reqwest_client: reqwest::Client,
+}
